requirements.txt
PyPDF2
pdfminer.six
pdfplumber
Pillow
pytesseract
pdf2image
tqdm